pyproject.toml
[project]
name = "first500daysassignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-documentintelligence>=1.0.2",
    "azure-identity>=1.25.1",
    "azure-search-documents>=11.6.0",
    "dotenv>=0.9.9",
    "fastapi>=0.128.0",
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-pinecone>=0.2.13",
    "langgraph>=1.0.5",
    "matplotlib>=3.10.8",
    "pinecone-text>=0.11.0",
    "seaborn>=0.13.2",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
